refactor(telegram): make code very clear to understand
This diff splits the telegram experiment such that each file
contain code related to a specific topic.

I'm ~already eying the possibility of making a tutorial out
of this codebase, so I'm trying to make it super clear.

While there, deal with recent telegram experiment changes
that caused the http://web.telegram.org/ checks to be dropped
as well as relaxing requirements on the returned webpage.