refactor: define category using >=> and return
This diff refactors our category definition using >=> (aka the
fish operator) and `return`. This definition is simpler than
the previous one and is the one used by "Category theory for
Computer Programmers" to introduce monads.

While there, recognize the previous code was not satisfying the
monadic laws and introduce Unit specifically for this purpose.

Users of this package will need to mechanically replace
invocations of ComposeFlatX with ComposeResultX.