refactor(dslx): compose observations using the monad
To implement this feature, I needed to redefine what monadic
composition is such that we can also merge observations as we
make our way forward through the composed functions.

It felt a bit weird to have this specific form of composition
inside fx, so I moved the new functionality back in dslx.