refactor(signal): use channel to await for background measurements
While the existing code is technically correct, we can simplify it by
just using the channel to await for background measurements.

This cuts in half the cognitive complexity of reading the code.

While there, also separate collecting all results from deciding whether
the overall result was a failure. While this change would reduce the
performance for large vectors, here we're dealing with a small one so
it's completely okay to refactor the code as such.